Improved Error Handling
Yamllint writes errors to stdout instead of stderr. I included a check
for the return value of the linting command. Both JSON and YAML tools
return a non-zero value, if something goes wrong during linting. This is
a more reliable check than the presence of content in stderr alone.
Let's hope other tools respect that as well.